+++ src/utils/form-element.ts
@@ -22,6 +22,7 @@ import {
   LocalFormElement,
   ShowIfHelper,
 } from '../types';
+import { createExecutionCode } from './code';
 import { disableIfCodeParameters, getOptionsCodeParameters, showIfCodeParameters } from './code-parameters';
 import { getFieldValues } from './grafana';
 
@@ -149,6 +150,7 @@ export const getElementWithNewType = (
         ...baseValues,
         value: '',
         type: newType,
+        isUseLocalTime: false,
       };
     }
     case FormElementType.BOOLEAN: {
@@ -238,10 +240,10 @@ export const isSectionCollisionExists = (sections: LayoutSection[], compareWith:
  */
 export const toLocalFormElement = (element: FormElement): LocalFormElement => {
   const showIf = element.showIf;
-
   let showIfFn: ShowIfHelper = () => true;
   if (showIf || showIf?.trim()) {
-    const fn = new Function('context', showIf);
+    const fn = createExecutionCode('context', showIf);
+
     showIfFn = ({ elements, replaceVariables }: { elements: FormElement[]; replaceVariables: InterpolateFunction }) =>
       fn(
         showIfCodeParameters.create({
@@ -259,7 +261,8 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
 
   let disableIfFn: DisableIfHelper = () => false;
   if (disableIf || disableIf?.trim()) {
-    const fn = new Function('context', disableIf);
+    const fn = createExecutionCode('context', disableIf);
+
     disableIfFn = ({
       elements,
       replaceVariables,
@@ -312,7 +315,8 @@ export const toLocalFormElement = (element: FormElement): LocalFormElement => {
         }));
       };
     } else if (element.optionsSource === OptionsSource.CODE) {
-      const fn = new Function('context', element.getOptions || 'return []');
+      const fn = createExecutionCode('context', element.getOptions || 'return []');
+
       getOptions = ({
         data,
         elements,
@@ -549,7 +553,7 @@ export const formatElementValue = (element: LocalFormElement, value: unknown): s
       return '*********';
     }
     case FormElementType.DATETIME: {
-      return value && typeof value === 'string' ? dateTime(value).toISOString() : '';
+      return value && typeof value === 'string' ? dateTime(value).toISOString(element.isUseLocalTime) : '';
     }
     case FormElementType.TIME: {
       return value && typeof value === 'string'