+++ CHANGELOG.md
@@ -30,8 +30,12 @@
 ### Features / Enhancements
 
 - Updated name to Business Forms Panel (#361)
-- Add support frontend data sources (#361)
-- Prepare for Grafana 11 (#399)
+- Added support frontend data sources (#361)
+- Prepared for Grafana 11 (#399)
+- Updated Reset button handler (#422)
+- Updated Confirmation Window (#420)
+- Updated Error handling in code editors showIf, disableIf, getOptions (#410)
+- Updated description for Get Options Code (#404)
 
 ## 3.8.0 (2024-05-30)